Add sweep support to RPC
This commit adds the ability to sweep an address for any and all
assets given a private key and an (optional) asset filter.

sweep privkey (asset_name)

The default is to sweep all assets, using the RVN hosted at the
swept account to pay all network fees. If the amount of RVN is
insufficient, then the remaining fee is paid by the balance
of the receiving wallet.

This feature can be tested by running `test/functional/feature_sweep.py`
or by runnning the entire `test_runner` testing framework.